feat: Add nuxt-security module and configure CORS handling
This commit adds the nuxt-security module to the nuxt.config.ts file and updates the package.json file to include the necessary dependencies. Additionally, it configures the CORS handling by allowing requests from any origin.

Refactor the code to improve performance

This commit applies lazy loading for images in the app.vue file, improving the performance of the application by loading images only when necessary. It also capitalizes the description in the AppHeading.vue component for consistency.

Update orderBy clause to order weekMenus by ID in descending order

This commit updates the orderBy clause in the current-menu.get.ts file to order the weekMenus by ID in descending order. This ensures that the weekMenus are displayed in the correct order.